requirements.txt
# gisfire-meteocat-lib @ file:///home/jaume/codi/tesi/MeteoCatLib/dist/gisfire_meteocat_lib-0.1.0.tar.gz

python-dateutil >= 2.8.2
pytz >= 2023.3
requests >= 2.31.0
SQLAlchemy >= 2.0.20
pytest >= 7.4.1
meteocat>=0.2.0
freezegun>=1.2.2